Mandrake v0.0.1 Mandrake.Annotation

Mandrake annotations.

Functions

deprecation(date_time, description)

Raise an exception if time to remove deprecation is over.

Examples

iex>  Mandrake.Annotation.deprecation({{2015, 12, 31}, {0, 0, 0}}, "Code refactoring needed")
nil

documentation(date_time, description)

Raise an exception if time to write documentation is over.

Examples

iex>  Mandrake.Annotation.documentation({{2015, 12, 31}, {0, 0, 0}}, "Write documentation for Math module")
nil

fix(date_time, description)

Raise an exception if time to fix an issue is over.

Examples

iex>  Mandrake.Annotation.fix({{2015, 12, 31}, {0, 0, 0}}, "Fix issue in 'sum' function")
nil

redundancy(date_time, description)

Raise an exception if time to fix redundancy is over.

Examples

iex>  Mandrake.Annotation.redundancy({{2015, 12, 31}, {0, 0, 0}}, "Fix redundancy")
nil

todo(date_time, description)

Raise an exception if time for todo is over.

Examples

iex>  Mandrake.Annotation.todo({{2015, 12, 31}, {0, 0, 0}}, "Must implement 'sum' function")
nil