Mandrake v0.0.4 Mandrake.Relation

Mandrake relational functions.

Functions

equals(value_1, value_2)

Returns true if the given values are equal.

Examples

iex>  Mandrake.Relation.equals('0', 0)
false

equals_by(function, value_1, value_2)

Returns true if the values produce the same result when passed to the provided function.

Examples

iex>  Mandrake.Relation.equals_by(fn arg -> Mandrake.Math.abs(arg) end, [-3], [3])
true

gt(first_value, second_value)

Returns true if first value is greater than the second.

Examples

iex>  Mandrake.Relation.gt(7, 5)
true

gte(first_value, second_value)

Returns true if first value is greater than or equal to the second.

Examples

iex>  Mandrake.Relation.gte(7, 7)
true

lt(first_value, second_value)

Returns true if first value is less than the second.

Examples

iex>  Mandrake.Relation.lt(7, 5)
false

lte(first_value, second_value)

Returns true if first value is less than or equal to the second.

Examples

iex>  Mandrake.Relation.lte(7, 7)
true

max(value_1, value_2)

Returns the larger of two arguments.

Examples

iex>  Mandrake.Relation.max('a', 'z')
'z'

max_by(function, value_1, value_2)

Returns the value that produces the larger result when passed to the provided function.

Examples

iex>  Mandrake.Relation.max_by(fn arg -> Mandrake.Math.negate(arg) end, -3, -6)
-6

min(value_1, value_2)

Returns the smaller of two arguments.

Examples

iex>  Mandrake.Relation.min('a', 'z')
'a'

min_by(function, value_1, value_2)

Returns the value that produces the smaller result when passed to the provided function.

Examples

iex>  Mandrake.Relation.min_by(fn arg -> Mandrake.Math.negate(arg) end, -3, -6)
-6