Mandrake v0.0.4 Mandrake.Type

Mandrake functions for types.

Functions

assign(first_tuple, second_tuple)

Returns the extension of the tuple with the second.

Examples

iex>  Mandrake.Type.assign({ "prop1", "prop2" }, { "prop3" })
{ "prop1", "prop2", "prop3" }

bool_false()

Always return false.

Examples

iex>  Mandrake.Type.bool_false()
false

bool_true()

Always return true.

Examples

iex>  Mandrake.Type.bool_true()
true

clone(object)

Create a copy of a value.

Examples

iex>  Mandrake.Type.clone([1, 2, 3, 4, 5])
[1, 2, 3, 4, 5]

exists(value)

Returns true if value is not nil.

Examples

iex>  Mandrake.Type.exists("Example")
true
iex>  Mandrake.Type.exists(nil)
false

falsy(value)

Returns true if value is nil or false.

Examples

iex>  Mandrake.Type.falsy("Example")
false
iex>  Mandrake.Type.falsy(nil)
true

is(type, value)

Returns true if the type of the value is the given type.

Examples

iex>  Mandrake.Type.is(Integer, 5)
true
iex>  Mandrake.Type.is(List, 5)
false

is_nil(value)

Returns true if value is nil.

Examples

iex>  Mandrake.Type.is_nil([])
false
iex>  Mandrake.Type.is_nil(nil)
true

split(pattern)

Returns a function that split a string with the given pattern.

Examples

iex>  splitComma = Mandrake.Type.split(",")
...>  splitComma.("a,b,c")
["a", "b", "c"]

trim(string)

Removes whitespace at the beginning and at the end of the string.

Examples

iex>  Mandrake.Type.trim("   Example   ")
"Example"

truthy(value)

Returns true if value is not nil or false.

Examples

iex>  Mandrake.Type.truthy("Example")
true
iex>  Mandrake.Type.truthy(nil)
false

type_of(value)

Returns type of value.

Examples

iex>  Mandrake.Type.type_of(42.0)
Float