Mandrake v0.0.1 Mandrake.Relation

Mandrake relational functions.

Functions

gt(first_value, second_value)

Returns true if first value is greater than the second.

Examples

iex>  Mandrake.Relation.gt(7, 5)
true

gte(first_value, second_value)

Returns true if first value is greater than or equal to the second.

Examples

iex>  Mandrake.Relation.gte(7, 7)
true

lt(first_value, second_value)

Returns true if first value is less than the second.

Examples

iex>  Mandrake.Relation.lt(7, 5)
false

lte(first_value, second_value)

Returns true if first value is less than or equal to the second.

Examples

iex>  Mandrake.Relation.lte(7, 7)
true